Project- Brain Tumor Detection app/Brain Tumor Detection/requirements.txt
flask
tensorflow
gdown
numpy
Pillow
matplotlib

